+++ .env.example
@@ -22,9 +22,9 @@ DATABASE_URL='mysql://o6te6c7ym1n611rrysd6:pscale_pw_QVonleCPhwUFQJtJ3gMsar3Rqio
 # You can generate a new secret on the command line with:
 # openssl rand -base64 32
 # https://next-auth.js.org/configuration/options#secret
-NEXTAUTH_SECRET="/4gwNHCR5872egAzn/dWp1tbGy8QI+F5fYS+8hBUvfc="
+NEXTAUTH_SECRET=""
 NEXTAUTH_URL="http://localhost:3000"
 
 # Next Auth Google Provider
-GOOGLE_CLIENT_ID=254640387206-9mov2tan6ff91npuj188ingpi820lrvb.apps.googleusercontent.com
-GOOGLE_CLIENT_SECRET=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
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
